Replace rules_docker with rules_oci
Because rules_grype expects the tarball to end in a `.tar`
suffix and otherwise assumes the base image has been provided,
`oci_tarball` targets were created for each image with names
corresponding to the implicit `.tar` rule previously provided by
rules_docker. Otherwise, `rules_oci` was able to be used as a
drop-in replacement.