Add support for MacOS arm64 build targets
The upstream version of rules_oci has a known bug that interferes with
running builds on MacOS with apple silicon because the jq binary does
not match the platform [1]. We patch with @archen's fork, which switches
that dependency to `yq`, which has platform-specific support.

Additionally, the latest release of `rules_anchore` has bugs that cause
builds to fail on platforms other than linux_amd64 [2]. We switch to the
latest patch, which contains a fix for that issue. In doing so, the
version of the grype vulnerability database was no longer compatible
with the version of grype in use, so we updated the grype database to
the latests version associated with the grype version targeted.

1: https://github.com/bazel-contrib/rules_oci/issues/253
2: https://github.com/hxtk/rules_anchore/issues/12