Update Go tooling to to latest version
io_bazel_rules_go contained incompatible changes at release 0.41.0 [1],
which had to be accounted for when upgrading to 0.42.0, along with
upgrading gazelle to 0.34.0. With these updates, we support Go 1.21.x,
so the Go toolchain itself was updated to 1.21.4.

The grpc_gateway code was removed because it is currently unused and was
not trivial to support with this toolchain upgrade.

In addition, gazelle-update-repos was run to rebuild repo definitions.

1: https://github.com/bazelbuild/rules_go/releases/tag/v0.41.0

TODO: restore grpc_gateway code.