Implement MapReduce
Define a Map worker capable of accepting records from an arbitrary
source (using a user-defined record stream), processing those logs
in a way consistent with the MapReduce API [1], and wiriting those
logs to a common intermediate format in sorted chunks decided by a
partition function.

Define a Reduce worker capable of reading records from the common
intermediate format and reducing those records in a manner consistent
with the reduce portion of the MapReduce API. These records are
processed in sorted order by merging arbitrarily many sorted chunks
as provided by the Map worker. The results are then committed to an
arbitrary destination, one per Reduce worker, using a user-defined
output record stream.

Implement the marshaling of the common intermediate record format
in a way that is consistent with the API of the user-defined input
and output record streams.

Implement a local API equivalent to GFS RecordAppend for the atomic
manipulation of Write-Ahead Log files by several processes.

TODO: Implement a MapReduce master to coordinate these jobs.